refactor: split VS Code webview assets
