+++ apps/vscode/media/main.js
@@ -1,0 +1,32 @@
+      const vscode = acquireVsCodeApi(); const messages = document.getElementById('messages'); const input = document.getElementById('input'); const status = document.getElementById('status'); const patchState = document.getElementById('patchState'); const patchPanel = document.getElementById('patchPanel'); const patchSummary = document.getElementById('patchSummary'); const patchFiles = document.getElementById('patchFiles'); const contextPanel = document.getElementById('contextPanel'); const contextSummary = document.getElementById('contextSummary'); const modeSelect = document.getElementById('mode'); let partial;
+      function escapeHtml(text){ return (text||'').replace(/[&<>]/g, ch => ({'&':'&amp;','<':'&lt;','>':'&gt;'}[ch])); }
+      function renderMarkdown(text){ const tick=String.fromCharCode(96); const fence=tick+tick+tick; const parts = String(text||'').split(new RegExp(fence+'([\\w.+-]*)\\n([\\s\\S]*?)'+fence,'g')); let html=''; for(let i=0;i<parts.length;i++){ if(i%3===0){ html += escapeHtml(parts[i]).replace(/^### (.*)$/gm,'<h3>$1</h3>').replace(/^## (.*)$/gm,'<h2>$1</h2>').replace(/^# (.*)$/gm,'<h1>$1</h1>').replace(new RegExp(tick+'([^'+tick+']+)'+tick,'g'),'<code>$1</code>').replace(/\n/g,'<br>'); } else if(i%3===1){ const lang=escapeHtml(parts[i]||'text'); const code=escapeHtml(parts[i+1]||''); html += '<pre><button class="copy" data-copy="code">Copy</button><code data-lang="'+lang+'">'+code+'</code></pre>'; i++; } } return html; }
+      function roleLabel(cls){ if(cls==='user') return 'You'; if(cls==='assistant') return 'Hermes'; if(cls==='error') return 'Error'; return 'Event'; }
+      function addAssistantActions(div){ const actions=document.createElement('div'); actions.className='msg-actions'; actions.innerHTML='<button class="secondary" data-copy="message">Copy</button><button class="secondary" data-retry="1">Retry</button>'; div.appendChild(actions); }
+      function add(cls, text){ const div=document.createElement('div'); div.className='msg '+cls; div.dataset.role=roleLabel(cls); div.innerHTML=renderMarkdown(text); if(cls==='assistant') addAssistantActions(div); messages.appendChild(div); messages.scrollTop=messages.scrollHeight; return div; }
+      function addTool(text){ const details=document.createElement('details'); details.className='tool-event'; details.innerHTML='<summary>🔧 '+escapeHtml(String(text).split('\n')[0]||'Hermes tool')+'</summary><div>'+renderMarkdown(text)+'</div>'; messages.appendChild(details); messages.scrollTop=messages.scrollHeight; }
+      function options(){ return {mode: modeSelect.value, includeFile:ctxFile.checked, includeSelection:ctxSel.checked, includeDiagnostics:ctxDiag.checked, includeGitDiff:ctxGit.checked, includeInstructions:ctxInst.checked}; }
+      function send(textOverride){ const text=textOverride || input.value; if(!text.trim()) return; input.value=''; vscode.postMessage({type:'ask', text, options: options()}); }
+      function setPatchState(value){ patchState.textContent=value; patchPanel.classList.toggle('visible', value && !/^No patch/.test(value)); }
+      function renderPatchDetails(value){ if(!value || !value.files || !value.files.length){ patchSummary.textContent='No patch metadata available.'; patchFiles.innerHTML=''; return; } patchSummary.textContent=value.files.length+' file(s), '+value.hunks+' hunk(s), +'+value.additions+'/-'+value.deletions; patchFiles.innerHTML=value.files.map(f=>'<div class="file-row"><span>'+escapeHtml(f.path||'unknown')+'</span><span><span class="delta-plus">+'+f.additions+'</span> <span class="delta-minus">-'+f.deletions+'</span> · '+f.hunks+' h</span></div>').join(''); }
+      function renderContextSummary(value){ if(!value){ contextSummary.textContent='No context summary available.'; return; } const included=Object.entries(value.included||{}).filter(([,on])=>on).map(([name])=>name).join(', ') || 'none'; contextSummary.innerHTML='<div><b>Workspace</b>: '+escapeHtml(value.workspace||'none')+'</div><div><b>Active file</b>: '+escapeHtml(value.activeFile||'none')+(value.dirty?' <em>(unsaved)</em>':'')+'</div><div><b>Language</b>: '+escapeHtml(value.language||'n/a')+'</div><div><b>Selection</b>: '+value.selectionLines+' line(s), '+value.selectionChars+' chars</div><div><b>Diagnostics</b>: '+value.diagnostics+'</div><div><b>Git status files</b>: '+value.gitFiles+'</div><div><b>Instructions</b>: '+escapeHtml((value.instructionFiles||[]).join(', ')||'none')+'</div><div><b>Included</b>: '+escapeHtml(included)+'</div><div><b>Budget</b>: '+value.maxContextChars+' chars</div>'; contextPanel.classList.add('visible'); }
+      document.getElementById('send').onclick=()=>send();
+      document.getElementById('review').onclick=()=>vscode.postMessage({type:'reviewDiff'});
+      document.getElementById('diag').onclick=()=>vscode.postMessage({type:'diagnostics'});
+      document.getElementById('term').onclick=()=>vscode.postMessage({type:'terminal'});
+      document.getElementById('context').onclick=()=>vscode.postMessage({type:'inspectContext', options: options()});
+      document.getElementById('preview').onclick=()=>vscode.postMessage({type:'previewPatch'});
+      document.getElementById('apply').onclick=()=>vscode.postMessage({type:'applyPatch'});
+      document.getElementById('copy').onclick=()=>vscode.postMessage({type:'copyPatch'});
+      document.getElementById('discard').onclick=()=>vscode.postMessage({type:'discardPatch'});
+      document.getElementById('runTests').onclick=()=>vscode.postMessage({type:'runTests'});
+      document.getElementById('runAnalyze').onclick=()=>vscode.postMessage({type:'runTestsAndAnalyze'});
+      document.getElementById('revert').onclick=()=>vscode.postMessage({type:'revertPatch'});
+      document.getElementById('stop').onclick=()=>vscode.postMessage({type:'stop'});
+      document.getElementById('clear').onclick=()=>vscode.postMessage({type:'clear'});
+      document.getElementById('config').onclick=()=>vscode.postMessage({type:'configure'});
+      input.addEventListener('keydown', e=>{ if ((e.metaKey||e.ctrlKey) && e.key === 'Enter') send(); });
+      document.addEventListener('click', e=>{ const prompt=e.target.dataset.prompt; if(prompt) send(prompt); if(e.target.dataset.copy==='message'){ navigator.clipboard.writeText(e.target.closest('.msg').innerText.replace(/^Hermes\n/,'').replace(/Copy\s*Retry$/,'')); } if(e.target.dataset.copy==='code'){ navigator.clipboard.writeText(e.target.parentElement.querySelector('code').innerText); } if(e.target.dataset.retry){ const last=[...messages.querySelectorAll('.user')].pop(); if(last) send(last.innerText.replace(/^You\n/,'')); } });
+      window.addEventListener('message', event=>{ const {type,value}=event.data; if(type==='restore'){ messages.innerHTML=''; if(!value.length) add('assistant','Ask Hermes to inspect, explain, edit, or review this workspace.'); value.forEach(m=>add(m.role === 'error' ? 'error' : m.role, m.text)); } if(type==='patchState') setPatchState(value); if(type==='patchDetails') renderPatchDetails(value); if(type==='contextSummary') renderContextSummary(value); if(type==='status') status.textContent=value; if(type==='user') { partial=undefined; add('user', value); } if(type==='partial') { if(!partial) partial=add('assistant', value); else partial.innerHTML=renderMarkdown(value); } if(type==='assistant') { if(partial) { partial.innerHTML=renderMarkdown(value); addAssistantActions(partial); } else add('assistant', value); partial=undefined; } if(type==='error') add('error', value); if(type==='tool') addTool(value); if(type==='patch') { addTool(value+' Use the Patch review panel to preview/apply/copy/discard.'); setPatchState(value); } if(type==='postApply') addTool(value); if(type==='patchCleared') addTool(value); });
+      vscode.postMessage({type:'ready'});
+
