+++ apps/vscode/scripts/package-check.js
@@ -4,6 +4,9 @@ const path = require('path');
 const root = path.resolve(__dirname, '..');
 const pkg = JSON.parse(fs.readFileSync(path.join(root, 'package.json'), 'utf8'));
 const missing = [];
+for (const file of pkg.files || []) {
+  if (!fs.existsSync(path.join(root, file))) missing.push(file);
+}
 for (const container of pkg.contributes.viewsContainers.activitybar || []) {
   if (container.icon && !fs.existsSync(path.join(root, container.icon))) missing.push(container.icon);
 }
